Working on entity eating system. After it is up and running will probably implement how need levels affect total happiness and work speed, etc + thoughts

+++ Dwarf/ECS/Components/Calender.h
@@ -5,7 +5,7 @@ class Calender : public Component
 {
 public:
 
-	int minute = 0;
+	double minute = 0.0;
 	int hour = 0;
 	int day = 0;
 	int month = 0;
